Add GET endpoint to UserController for user retrieval
A new asynchronous HTTP GET endpoint has been added to the `UserController` class in the `RMS_API.Controllers.Users` namespace. The endpoint, defined with the route `"{id}"`, retrieves user information based on the provided user ID. It returns an `ActionResult` containing an `IEnumerable<UserModel>`. The method filters active users with the specified user ID and projects the data into a `UserModel` object. The result is returned as an `Ok` response, and a 500 Internal Server Error status code is returned if an exception occurs.